+++ lib/msf/core/encoder.rb
@@ -684,4 +684,5 @@ require 'msf/core/encoder/xor_additive_feedback'
 require 'msf/core/encoder/alphanum'
 require 'msf/core/encoder/nonalpha'
 require 'msf/core/encoder/nonupper'
+require 'msf/core/encoder/xor_dynamic'
