+++ modules/encoders/x86/xor_dynamic.rb
@@ -3,7 +3,7 @@
 # Current source: https://github.com/rapid7/metasploit-framework
 ##
 
-class MetasploitModule < Msf::Encoder
+class MetasploitModule < Msf::Encoder::XorDynamic
 
   def initialize
     super(
@@ -11,176 +11,40 @@ class MetasploitModule < Msf::Encoder
       'Description'      => 'An x86 XOR encoder with dynamic key size',
       'Author'           => [ 'lupman', 'phra' ],
       'Arch'             => ARCH_X86,
-      'License'          => MSF_LICENSE,
-      'EncoderType'      => Msf::Encoder::Type::Raw
+      'License'          => MSF_LICENSE
       )
   end
 
-  def minKeyLen
-    Integer(datastore['KEYMIN'] || 0)
-  end
-
-  def maxKeyLen
-    Integer(datastore['KEYMAX'] || 0)
-  end
-
   def stub
-    # jmp short _call
-    # _ret:
-    #   pop ecx
-    #   mov ebx, ecx
-    #   mov esi, ebx
-    # lp1:
-    #   cmp byte [ecx], 'a'
-    #   je _ok
-    #   inc ecx
-    #   jmp lp1
-    # _ok:
-    #   inc ecx
-    #   mov edi, ecx
-    # lp:
-    #   cmp word [ecx], 'bb'
-    #   jz _jmp
-    #   mov al, byte [ebx]
-    #   xor byte [ecx], al
-    #   inc ecx
-    #   inc ebx
-    #   cmp byte [ebx], 'a'
-    #   jnz lp
-    #   mov ebx, esi
-    #   jmp lp
-    # _jmp:
-    #   jmp edi
-    # _call:
-    #   call _ret
-    "\xeb\x28\x59\x89\xcb\x89\xde\x80\x39\x61\x74\x03\x41\xeb\xf8\x41\x89\xcf\x66\x81\x39\x62\x62\x74\x0f\x8a\x03\x30\x01\x41\x43\x80\x3b\x61\x75\xee\x89\xf3\xeb\xea\xff\xe7\xe8\xd3\xff\xff\xff"
+    "\xeb\x28" +              #         jmp   _call
+    "\x59" +                  # _ret:   pop   ecx
+    "\x89\xcb" +              #         mov   ebx, ecx
+    "\x89\xde" +              #         mov   esi, ebx
+    "\x80\x39\x61" +          # _lp1:   cmp   BYTE PTR [ecx], 'a'
+    "\x74\x03" +              #         je    _ok
+    "\x41" +                  #         inc   ecx
+    "\xeb\xf8" +              #         jmp   _lp1
+    "\x41" +                  # _ok:    inc   ecx
+    "\x89\xcf" +              #         mov   edi,ecx
+    "\x66\x81\x39\x62\x62" +  # _lp:    cmp   WORD PTR [ecx], 'bb'
+    "\x74\x0f" +              #         je    _jmp
+    "\x8a\x03" +              #         mov   al,BYTE PTR [ebx]
+    "\x30\x01" +              #         xor   BYTE PTR [ecx], al
+    "\x41" +                  #         inc   ecx
+    "\x43" +                  #         inc   ebx
+    "\x80\x3b\x61" +          #         cmp   BYTE PTR [ebx], 'a'
+    "\x75\xee" +              #         jne   _lp
+    "\x89\xf3" +              #         mov   ebx, esi
+    "\xeb\xea" +              #         jmp   _lp
+    "\xff\xe7" +              # _jmp:   jmp   edi
+    "\xe8\xd3\xff\xff\xff"    # _call:  call  _ret
   end
 
-  def find_key(buf, badchars, keyChars)
-
-    keyFound = nil
-
-    bufLen = buf.length
-
-    # Search for a valid key
-    _minKeyLen = minKeyLen
-    if _minKeyLen < 1
-      _minKeyLen = Integer(buf.length / 100 * (0.2 + 0.05 * badchars.length))
-      if _minKeyLen < 1
-        _minKeyLen = 1
-      end
-    end
-
-    _maxKeyLen = maxKeyLen
-    if _maxKeyLen < 1
-      _maxKeyLen = buf.length
-    end
-
-    for keyLen in _minKeyLen.._maxKeyLen do
-      $stderr.print "\rKey size: #{keyLen}"
-      $stderr.flush
-
-      myKey = ""
-      for x in 0..keyLen - 1 do
-        keyChars.each_char do |j|
-          ok = true
-          i = 0
-          while i + x < bufLen do
-            if badchars[(buf[i + x].ord ^ j.ord).chr]
-              ok = false
-              break
-            end
-
-            i += keyLen
-          end
-
-          if ok
-            myKey << j.chr
-            break
-          end
-
-        end
-      end
-
-      if myKey.length == keyLen
-        keyFound = myKey
-        break
-      end
-    end
-
-    $stderr.print "\n"
-    $stderr.flush
-    return keyFound
+  def stubKeyTerm
+    /a/
   end
 
-  def encode(buf, badchars = nil, state = nil, platform = nil)
-
-    # Set default badchars if empty
-    badchars = "\x00\x0a\x0d" if (badchars == nil or badchars == '')
-
-    # Check badchars in stub
-    if Rex::Text.badchar_index(stub.gsub(/A/, "").gsub(/BB/, ""), badchars)
-      raise EncodingError, "Bad character found in stub for the #{self.name} encoder.", caller
-    end
-
-    # Set allowed chars
-    keyChars = ""
-    for i in 1..255 do
-      if !badchars[i.chr]
-        keyChars << i.chr
-      end
-    end
-
-    # Find key
-    key = find_key(buf, badchars, keyChars)
-
-    if key == nil
-      raise NoKeyError, "A key could not be found for the #{self.name} encoder.", caller
-    end
-
-    # Search for key terminator
-    keyTerm = nil
-    keyChars.chars.shuffle.each do |i|
-      if !key[i]
-        keyTerm = i
-        break
-      end
-    end
-
-    if keyTerm == nil
-      raise EncodingError, "Key terminator could not be found for the #{self.name} encoder.", caller
-    end
-
-    # Encode paylod
-    pos = 0
-    encoded = ""
-    while pos < buf.length
-      encoded << (buf[pos].ord ^ key[pos % key.length].ord).chr
-      pos += 1
-    end
-
-    # Search for payload terminator
-    payloadTerm = nil
-    keyChars.chars.shuffle.each do |i|
-      break unless keyChars.chars.shuffle.each do |j|
-        if !encoded.index(i + j)
-          payloadTerm = i + j
-          break
-        end
-      end
-    end
-
-    if payloadTerm == nil
-      raise EncodingError, "Payload terminator could not be found for the #{self.name} encoder.", caller
-    end
-
-    finalPayload = stub.gsub(/a/, keyTerm).gsub(/bb/, payloadTerm) + key + keyTerm + encoded + payloadTerm
-
-    # Check badchars in finalPayload
-    if Rex::Text.badchar_index(finalPayload, badchars)
-     raise EncodingError, "Bad character found for the #{self.name} encoder.", caller
-    end
-
-    return finalPayload
+  def stubPayloadTerm
+    /bb/
   end
 end
